requirements.txt
absl-py==0.9.0
APScheduler==3.6.3
asgiref==3.2.3
astor==0.8.1
astroid==2.4.2
bottle==0.12.18
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.2
django-cors-headers==3.4.0
djangorestframework==3.11.0
gast==0.2.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
isort==4.3.21
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
marshmallow==3.5.1
mccabe==0.6.1
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
packaging==19.2
platformio==4.3.4
protobuf==3.11.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyelftools==0.26
pylint==2.5.3
pyparsing==2.4.5
PyQt5==5.13.2
PyQt5-sip==12.7.0
pyserial==3.4
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
semantic-version==2.8.4
sip==5.0.0
six==1.14.0
sqlparse==0.3.0
tabulate==0.8.6
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
toml==0.10.0
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.8
Werkzeug==1.0.0
whitenoise==5.1.0
wrapt==1.12.1
